cleans up fixtures
