+++ cli/tests/fixtures/rich/output.txt
